feat(tap-on-notification): Handled notifications that imply refreshing the applets
