+++ src/entities/applet/model/hooks/index.ts
@@ -1,3 +1,3 @@
 export { default as useInProgressRecord } from './useInProgressRecord';
-
 export { default as useStartEntity } from './useStartEntity';
+export { default as useRefreshMutation } from './useRefreshMutation';
