+++ src/entities/applet/model/hooks/useRefreshMutation.ts
@@ -1,0 +1,26 @@
+import { useMemo } from 'react';
+
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+
+import { RefreshService } from '../services';
+
+const useRefreshMutation = (onSuccess: () => void) => {
+  const queryClient = useQueryClient();
+
+  const refreshService = useMemo(
+    () => new RefreshService(queryClient),
+    [queryClient],
+  );
+
+  const refresh = useMemo(
+    () => refreshService.refresh.bind(refreshService),
+    [refreshService],
+  );
+
+  return useMutation(['refresh'], refresh, {
+    networkMode: 'always',
+    onSuccess,
+  });
+};
+
+export default useRefreshMutation;
