+++ src/entities/applet/model/services/index.ts
@@ -1,0 +1,1 @@
+export { default as RefreshService } from './RefreshService';
