+++ src/features/applets-refresh/model/hooks/useAutomaticRefreshOnMount.ts
@@ -3,12 +3,11 @@ import { useEffect } from 'react';
 import { useIsRestoring } from '@tanstack/react-query';
 
 import { useCacheHasData } from '@app/shared/lib';
-
-import useRefreshMutation from './useRefreshMutation';
+import { AppletModel } from '@entities/applet';
 
 function useAutomaticRefreshOnMount(onSuccess: () => void) {
   const isCacheRestoring = useIsRestoring();
-  const { mutate: refresh } = useRefreshMutation(onSuccess);
+  const { mutate: refresh } = AppletModel.useRefreshMutation(onSuccess);
   const { check: checkIfCacheHasData } = useCacheHasData();
 
   useEffect(() => {
