+++ src/features/applets-refresh/model/hooks/useRefresh.ts
@@ -1,9 +1,9 @@
 import { useState } from 'react';
 
-import useRefreshMutation from './useRefreshMutation';
+import { AppletModel } from '@entities/applet';
 
 function useRefresh(onSuccess: () => void) {
-  const { mutateAsync: refresh } = useRefreshMutation(onSuccess);
+  const { mutateAsync: refresh } = AppletModel.useRefreshMutation(onSuccess);
 
   const [isRefreshing, setIsRefreshing] = useState(false);
